Break up dashboard into components
